Aula 16/12/2020

+++ backend/controllers/AuthorController.php
@@ -7,6 +7,7 @@ use common\models\Author;
 use yii\data\ActiveDataProvider;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\HttpException;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
 
@@ -23,7 +24,6 @@ class AuthorController extends Controller
         return [
             'access' => [
                 'class' => AccessControl::className(),
-                'only' => ['list', 'view', 'create', 'update', 'delete'],
                 'rules' => [
                     [
                         'allow' => true,
@@ -33,22 +33,17 @@ class AuthorController extends Controller
                     [
                         'allow' => true,
                         'actions' => ['view'],
-                        'roles' => ['ViewPost'],
+                        'roles' => ['admin','full_manager','medium_manager','low_manager'],
                     ],
                     [
                         'allow' => true,
                         'actions' => ['create'],
-                        'roles' => ['CreatePost'],
+                        'roles' => ['admin','full_manager','medium_manager'],
                     ],
                     [
                         'allow' => true,
-                        'actions' => ['update'],
-                        'roles' => ['UpdatePost'],
-                    ],
-                    [
-                        'allow' => true,
-                        'actions' => ['delete'],
-                        'roles' => ['DeletePost'],
+                        'actions' => ['update','delete'],
+                        'roles' => ['admin','full_manager'],
                     ],
                 ],
             ],
@@ -67,6 +62,10 @@ class AuthorController extends Controller
      */
     public function actionList()
     {
+        if(!Yii::$app->user->can('ViewPost')){
+            throw new HttpException(403,'You are not allowed to perform this action.');
+        }
+
         $Authors = Author::find()->all();
 
         return $this->render('list', [
@@ -82,6 +81,10 @@ class AuthorController extends Controller
      */
     public function actionView($idAuthor)
     {
+        if(!Yii::$app->user->can('ViewPost')){
+            throw new HttpException(403,'You are not allowed to perform this action.');
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($idAuthor),
         ]);
@@ -94,6 +97,10 @@ class AuthorController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can('CreatePost')){
+            throw new HttpException(403,'You are not allowed to perform this action.');
+        }
+
         $model = new Author();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -114,6 +121,10 @@ class AuthorController extends Controller
      */
     public function actionUpdate($idAuthor)
     {
+        if(!Yii::$app->user->can('UpdatePost')){
+            throw new HttpException(403,'You are not allowed to perform this action.');
+        }
+
         $model = $this->findModel($idAuthor);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -134,6 +145,10 @@ class AuthorController extends Controller
      */
     public function actionDelete($idAuthor)
     {
+        if(!Yii::$app->user->can('DeletePost')){
+            throw new HttpException(403,'You are not allowed to perform this action.');
+        }
+
         $this->findModel($idAuthor)->delete();
 
         return $this->redirect(Yii::$app->request->baseUrl.'/author_list');
